Add support for async (create/delete/actions) floating_ips requests.
Other async requests in godotto follow a pattern of issuing the action then polling the associated /resource/:id/actions endpoint to wait for the completion of the action.
For example Droplet Create does just that. The floating IP assign/unassign actions do not currently have the wait support so assertions against their state are flappy.